+++ module/Soutenance/src/Soutenance/Controller/JustificatifController.php
@@ -125,25 +125,30 @@ class JustificatifController extends AbstractController
         return $vm;
     }
 
-    public function ajouterAutorisationSoutenanceAction() : ViewModel
+    public function ajouterDocumentLieSoutenanceAction() : ViewModel
     {
+        $codeNatureFichier = $this->params()->fromQuery('nature');
+        $labelFichier = $this->params()->fromQuery('label');
+
         $these = $this->requestedThese();
         $proposition = $this->getPropositionService()->findOneForThese($these);
 
         $fichier = new Fichier();
-        $nature = $this->fichierTheseService->fetchNatureFichier(NatureFichier::CODE_AUTORISATION_SOUTENANCE);
+        $nature = $this->fichierTheseService->fetchNatureFichier($codeNatureFichier);
         $fichier->setNature($nature);
 
         $justificatif = new Justificatif();
         $justificatif->setProposition($proposition);
         $form = $this->getJustificatifForm();
-        $form->setAttribute('action', $this->url()->fromRoute('soutenance/justificatif/ajouter-autorisation-soutenance', ['these' => $these->getId()], [], true));
+        $form->setAttribute('action', $this->url()->fromRoute('soutenance/justificatif/ajouter-document-lie-soutenance', ['these' => $these->getId()], ["query" => [
+            "nature" => $codeNatureFichier,
+            "label" => $labelFichier
+        ]], true));
         $form->bind($justificatif);
         $form->init();
 
         $request = $this->getRequest();
         if ($request->isPost()) {
-//            $data = $request->getPost();
             $files = ['files' => $request->getFiles()->toArray()];
 
             if (!empty($files)) {
@@ -157,88 +162,7 @@ class JustificatifController extends AbstractController
 
 
         $vm =  new ViewModel([
-            'title' => "Téléversement de l' autorisation de soutenance",
-            'these' => $these,
-            'form' => $form,
-        ]);
-        $vm->setTemplate('soutenance/justificatif/ajouter');
-        return $vm;
-    }
-
-
-    public function ajouterRapportSoutenanceAction() : ViewModel
-    {
-        $these = $this->requestedThese();
-        $proposition = $this->getPropositionService()->findOneForThese($these);
-
-        $fichier = new Fichier();
-        $nature = $this->fichierTheseService->fetchNatureFichier(NatureFichier::CODE_RAPPORT_SOUTENANCE);
-        $fichier->setNature($nature);
-
-        $justificatif = new Justificatif();
-        $justificatif->setProposition($proposition);
-        $form = $this->getJustificatifForm();
-        $form->setAttribute('action', $this->url()->fromRoute('soutenance/justificatif/ajouter-rapport-soutenance', ['these' => $these->getId()], [], true));
-        $form->bind($justificatif);
-        $form->init();
-
-        $request = $this->getRequest();
-        if ($request->isPost()) {
-//            $data = $request->getPost();
-            $files = ['files' => $request->getFiles()->toArray()];
-
-            if (!empty($files)) {
-                $version = $this->fichierTheseService->fetchVersionFichier(VersionFichier::CODE_ORIG);
-                $fichiers = $this->fichierTheseService->createFichierThesesFromUpload($these, $files, $nature, $version);
-                $justificatif->setFichier($fichiers[0]);
-                $this->getJustificatifService()->create($justificatif);
-            }
-            exit();
-        }
-
-
-        $vm =  new ViewModel([
-            'title' => "Téléversement du rapport de soutenance",
-            'these' => $these,
-            'form' => $form,
-        ]);
-        $vm->setTemplate('soutenance/justificatif/ajouter');
-        return $vm;
-    }
-
-    public function ajouterProcesVerbalSoutenanceAction() : ViewModel
-    {
-        $these = $this->requestedThese();
-        $proposition = $this->getPropositionService()->findOneForThese($these);
-
-        $fichier = new Fichier();
-        $nature = $this->fichierTheseService->fetchNatureFichier(NatureFichier::CODE_PV_SOUTENANCE);
-        $fichier->setNature($nature);
-
-        $justificatif = new Justificatif();
-        $justificatif->setProposition($proposition);
-        $form = $this->getJustificatifForm();
-        $form->setAttribute('action', $this->url()->fromRoute('soutenance/justificatif/ajouter-proces-verbal-soutenance', ['these' => $these->getId()], [], true));
-        $form->bind($justificatif);
-        $form->init();
-
-        $request = $this->getRequest();
-        if ($request->isPost()) {
-//            $data = $request->getPost();
-            $files = ['files' => $request->getFiles()->toArray()];
-
-            if (!empty($files)) {
-                $version = $this->fichierTheseService->fetchVersionFichier(VersionFichier::CODE_ORIG);
-                $fichiers = $this->fichierTheseService->createFichierThesesFromUpload($these, $files, $nature, $version);
-                $justificatif->setFichier($fichiers[0]);
-                $this->getJustificatifService()->create($justificatif);
-            }
-            exit();
-        }
-
-
-        $vm =  new ViewModel([
-            'title' => "Téléversement du procès-verbal de soutenance",
+            'title' => "Téléversement du document : ".$labelFichier,
             'these' => $these,
             'form' => $form,
         ]);